feat: 종목/지수 차트 상세 화면 + 시간별·봉차트 + 크래시 수정

+++ lib/features/market/presentation/providers/market_etf_ranking_controller.dart
@@ -49,7 +49,7 @@ class MarketEtfRankingController
               name: d.name,
               code: d.code,
               changePercent: d.changeRate,
-              price: d.price,
+              price: d.price.toDouble(),
             ))
         .toList();
     return MarketEtfRankingState(filter: filter, items: items);
@@ -80,7 +80,7 @@ class MarketUsEtfRankingController
               name: d.name,
               code: d.code,
               changePercent: d.changeRate,
-              price: d.price,
+              price: d.price.toDouble(),
             ))
         .toList();
     return MarketEtfRankingState(filter: filter, items: items);
